Update FilepondService.php for fixing chunk upload
The most important change is to verify that each file found is a valid chunk. You can't rely just on the filename.  A better approach is to use the Upload-Offset header to determine the expected offset of each chunk file and compare it to the filename.

Before attempting to fread, explicitly check the file size. This is the direct fix for the error.